+++ gatenlp/processing/gazetteer/stringregex.py
@@ -3,30 +3,50 @@ Module that defines classes for matching annotators other than gazetteers which 
 of strings or annotations.
 """
 import re
-from typing import Union, List, Set
+from typing import Union, List, Set, Optional, Dict, Iterable, Any, Tuple
 from collections import namedtuple
+from recordclass import structclass
 from gatenlp import Document
-from gatenlp.processing.gazetteer.base import StringGazetteerBase, Match
+from gatenlp.processing.gazetteer.base import StringGazetteerBase
+
+# TODO: use named tuple instead of tuple for each action, to make it easier to keep track of which field used?
+
 
 # rule body line:
 # one or more comma separated group numbers followed by a "=>" followed by feature assignments
 # each feature assignment is a name followed by "=" followed by some non whitespace
 # featuer assignments are comma-separated (whitespace can be added)
 # e.g. 0,3 => Lookup f1=2, f2 = "something" , f3=True
-PAT_RULE_BODY_LINE = re.compile(r"^\s*([0-9]+(?:,[0-9]+)*)\s*=>\s*(\w+)(\s*\w+\s*=\s*\S+)*\s*$")
-PAT_RULE_BODY_FASSIGN = re.compile(r"\s*(\w+)\s*=\s*(\S+)")
-PAT_DOLLARN = re.compile(r"^\$[0-9]+$")
+PAT_RULE_BODY_LINE = re.compile(r"^\s*([0-9]+(?:,[0-9]+)*)\s*=>\s*(\w+)\s*(.*)$")
 PAT_MACRO_LINE = re.compile(r"\s*([a-zA-Z0-9_]+)=(\S+)\s*$")
 PAT_SUBST = re.compile(r"{{\s*[a-zA-Z0-9_]+\s*}}")
 
 GroupNumber = namedtuple("GroupNumber", ["n"])
 
-def one_or(someiterator, default=None):
+GLOBALS = {f"G{i}": GroupNumber(i) for i in range(99)}
+
+Rule = namedtuple("Rule", ["pattern", "actions"])
+Action = namedtuple("Action", ["groupnumbers", "typename", "features"])
+
+Match = structclass(
+    # A description of a match.
+    # Fields:
+    # start: the offset or index (in case matching a token list) where the match starts
+    # end: the offset or index one past where the match ends
+    # match: the matched string
+    # features: the features as merged from the match rule/entry and the list features
+    # type : the type as specified in the rule
+    "Match", ("start", "end", "match", "features", "type")
+)
+
+
+def one_or(someiterator: Iterable, default: Any = None) -> Any:
     for el in someiterator:
         return el
     return default
 
-def subst(strg, substs):
+
+def subst(strg: str, substs: Optional[Dict]) -> str:
     """
     Substitute any predefined replacements in the strg. If the strg contains sometime {{name}} and
     "name" is a key in substs, the "{{name}}" substring gets replaced with the value in substs.
@@ -35,11 +55,13 @@ def subst(strg, substs):
 
     Args:
         strg: some string
-        substs: dictionary of name->replacementstring
+        substs: dictionary of name->replacementstring or None
 
     Returns:
         the modified string
     """
+    if substs is None or len(substs) == 0:
+        return strg
     matches = list(re.finditer(PAT_SUBST, strg))
     if len(matches) == 0:
         return strg
@@ -51,6 +73,7 @@ def subst(strg, substs):
         var = txt[2:-2]
         val = substs.get(var)
         if val is not None:
+            val = "(?:" + str(val) + ")"
             txt = subst(str(val), substs)
         if start > lastidx:
             parts.append(strg[lastidx:start])
@@ -62,49 +85,14 @@ def subst(strg, substs):
     return ret
 
 
-def make_rule(pats, acts, substs, backend="re"):
-    """
-    Create a rule representation from a list of pattern strings, a list of action line tuples
-    (groupnumber, typename, featureassignments) and a dictionary of name substitutions.
-
-    Args:
-        pats: list or pattern strings
-        acts: list of tuples (groupnumber, typename, featureassignments)
-        substs: dictionary of name -> string substituions
-
-    Returns:
-        a rule representation, a tuple where the first element is a compiled regexp and the second element
-            is a list of of annotation descriptions.
-            Each annotation description is a tuple with: list of group numbers, annotation type name,
-            dictionary of featurename to value mappings. If the value is an instance of GroupNumber, then
-            it represents the match group number to take the value of the feature from
-    """
-    pats = ["(?:"+subst(p.strip(), substs)+")" for p in pats]
-    # TODO: use the correct backend!
-    if backend == "re":
-        pattern = re.compile("".join(pats))
-    elif backend == "regex":
-        pattern = regex.compile("".join(pats))  # noqa: F821
-    else:
-        raise Exception(f"Invalid regular expression backend: {backend}")
-    anndescs = []
-    for act in acts:
-        grpnrs, typname, fstr = act
-        grpnrs = [int(x) for x in grpnrs.strip().split(",")]
-        fassignments = {}
-        for m in re.findall(PAT_RULE_BODY_FASSIGN, fstr):
-            fname, fval = m
-            # ignore commas
-            if fval.endswith(","):
-                fval = fval[:-1]
-            dollarn = re.match(PAT_DOLLARN, fval)
-            if dollarn:
-                fval = GroupNumber(int(fval[1:]))
-            else:
-                fval = eval(fval)
-            fassignments[fname] = fval
-        anndescs.append((grpnrs, typname, fassignments))
-    return pattern, anndescs
+def replace_group(feats: Dict, groups: Union[list, tuple]):
+    """Replace any group placeholders in any of the feats with the text from the group in the match"""
+    ret = {}
+    for k, v in feats.items():
+        if isinstance(v, GroupNumber):
+            v = groups[v.n]
+        ret[k] = v
+    return ret
 
 
 class StringRegexAnnotator(StringGazetteerBase):
@@ -118,21 +106,21 @@ class StringRegexAnnotator(StringGazetteerBase):
                  containing_type=None,
                  list_features=None,
                  skip_longest=False,
-                 match="all",
-                 engine='re'
+                 longest_only: bool = False,
+                 select_rules="all",  # or first / last
+                 regex_module='re',
+                 start_type: Optional[str] = None,
+                 end_type: Optional[str] = None,
+                 split_type: Optional[str] = None,
                  ):
         """
         Create a StringRegexAnnotator and optionally load regex patterns.
 
         Args:
             source: where to load the regex rules from, format depends on source_fmt. If None, nothing is loaded.
-            source_fmt: the format of the source. Either "list" for a list of tuples, where the first element
-                is a compiled regular expression and the second element is a list of annotation descriptions.
-                Each annotation description is a tuple with: list of group numbers, annotation type name,
-                dictionary of featurename to value mappings. If the value is an instance of GroupNumber, then
-                it represents the match group number to take the value of the feature from
-                Or the source_fmt can be "file" in which case the rules are loaded from a file with that path.
-                Or the source_fmt can be "string" in which case the rules are loaded from the given string.
+            source_fmt: the format of the source. Either "rule-list" for a list of Rule instances,
+                or "file" in which case the rules are loaded from a file with that path
+                or "string" in which case the rules are loaded from the given string.
             outset_name: name of the output annotation set
             annset_name: the input annotation set if matching is restricted to spans covered by
                 containing annotations.
@@ -143,54 +131,84 @@ class StringRegexAnnotator(StringGazetteerBase):
             skip_longest: if True, after a match, the next match is attempted after the longest current
                 match. If False, the next match is attempted at the next offset after the start of the
                 current match.
-            match: the strategy of which rule to apply. One of: "all": apply all matching rules.
-                "first": apply the first matching rule, do not attempt any others. "firstlongest":
-                try all rules, apply the first of all rules with the longest match. "alllongest":
-                try all rules, apply all rules with the longest match.
-            engine: identifies which Python regular expression engine to use. Currently either
-                "re" or "regexp" to use the package with the corresponding name. Only the package
+            longest_only: if True, only use the longest match(es) of all matches
+            select_rules: the strategy of which rules select from all that have acceptable matches.
+                One of: "all": apply all matching rules (all matching or all longest depending on longest_only)
+                "first": apply the first of all matching/first of all longest matching rules
+                "last": apply the last of all matching/all longest matching rules
+            regex_module: the name (string) of the regular expression module to use. Currently either
+                "re" or "regexp" to use the module with the corresponding name. Only the module
                 used is attempted to get loaded.
+            start_type: if not None, the annotation type of annotations defining possible starting points of matches,
+                if None, matches can occur anywhere
+            end_type: if not None, the annotation type of annotations defining possible end points of matches, if
+                None, matches can end anywhere
+            split_type: the annotation type of annotations indicating splits, no match can cross a split
         """
         self.rules = []
         self.outset_name = outset_name
         self.annset_name = annset_name
         self.containing_type = containing_type
         self.skip_longest = skip_longest
-        self.match = match
-        self.engine = engine
-        if engine == "regex":
+        self.longest_only = longest_only
+        self.select_rules = select_rules
+        self.regex_module = regex_module
+        self.start_type = start_type
+        self.end_type = end_type
+        self.split_type = split_type
+        self.list_features = list_features
+        if regex_module == "regex":
             try:
                 import regex
+                self.re = regex
             except Exception as ex:
-                raise Exception(f"Cannot use regex, import failed")
+                raise Exception(f"Cannot use module regex, import failed", ex)
+        else:
+            self.re = re
         self.features4list = []
         if source is not None:
             self.append(source, source_fmt=source_fmt, list_features=list_features)
 
-    def add(self, rule, list_features=None):
+    def make_rule(self,
+                  pats: Union[str, List[str]],
+                  acts: List[Tuple[str, str, dict]],
+                  substs: Optional[Dict] = None,
+                  list_features: Optional[Dict] = None,
+                  ):
         """
-        Add a single rule.
+        Create a rule representation from a pattern string or list of pattern strings, an a list of action line tuples
+        (groupnumbersstring, typenamestring, featureassignmentsstring) and a dictionary of name substitutions.
+
+        Each featureassignmentsstring is a string of the form "fname1=fval1, fname2=fval2".
 
         Args:
-            rule: a tuple, where the first element
-                is a compiled regular expression and the second element is a list of annotation descriptions.
-                Each annotation description is a tuple with: list of group numbers, annotation type name,
-                dictionary of featurename to value mappings. If the value is an instance of GroupNumber, then
-                it represents the match group number to take the value of the feature from
-            list_features: a dictionary of features to add to the rule
-        """
-        pat, anndescs = rule
-        if list_features is not None:
-            newdescs = []
-            for anndesc in anndescs:
-                grpnrs, typename, feats = anndesc
-                features = list_features.copy()
-                features.update(feats)
-                newdescs.append((grpnrs, typename, features))
-            anndescs = newdescs
-        self.rules.append((pat, anndescs))
+            pats: a list of pattern strings, or a single pattern string
+            acts: list of tuples (groupnumbersstring, typenamestring, features)
+            substs: dictionary of name -> string substituions
+            list_features: features to add to each action, if None, add the ones specified at init time
 
-    def append(self, source, source_fmt="file", list_features=None):
+        Returns:
+            a Rule instance
+        """
+        if list_features is None:
+            list_features = self.list_features
+            if list_features is None:
+                list_features = {}
+        if isinstance(pats, str):
+            pats = subst(pats.strip(), substs)
+        else:
+            pats = ["(?:" + subst(p.strip(), substs) + ")" for p in pats]
+        pattern = self.re.compile("".join(pats))  # noqa: F821
+        anndescs = []
+        for act in acts:
+            grpnrs, typname, feats = act
+            grpnrs = [int(x) for x in grpnrs.strip().split(",")]
+            fassignments = list_features.copy()
+            fassignments.update(feats)
+            anndescs.append(Action(groupnumbers=grpnrs, typename=typname, features=fassignments))
+        return Rule(pattern=pattern, actions=anndescs)
+
+    def append(self, source: Union[str, List], source_fmt: str = "file", list_features = None):
         """
         Load a list of rules.
 
@@ -200,15 +218,11 @@ class StringRegexAnnotator(StringGazetteerBase):
             list_features: if not None a dictionary of features to assign to annotations created
                 for any of the rules loaded by that method call.
         """
-        # Format of rule: a tuple where the first element is a compiled regular expression
-        # and the second element is a tuple. That tuple describes the
-        # annotations to create for a match. The first element of the tuple is the annotation type or None
-        # to use the out_type. The second element is a dictionary mapping each group number of the match
-        # to a dictionary of features to assign. If the feature value is the string "$n" with n a group
-        # number, then the value for that match group is used.
-        if source_fmt == "list":
+        if list_features is None:
+            list_features = self.list_features
+        if source_fmt == "rule-list":
             for rule in source:
-                self.add(rule, list_features=list_features)
+                self.rules.append(rule)
         else:
             if source_fmt == "file":
                 with open(source, "rt", encoding="utf-8") as infp:
@@ -231,16 +245,23 @@ class StringRegexAnnotator(StringGazetteerBase):
                     # if there is a current rule, add it
                     if len(cur_acts) > 0:
                         # finish and add rule
-                        rule = make_rule(cur_pats, cur_acts, cur_substs, backend=self.engine)
+                        rule = self.make_rule(cur_pats, cur_acts, substs=cur_substs, list_features=list_features)
                         cur_acts = []
-                        self.add(rule, list_features=list_features)
+                        cur_pats = []
+                        self.rules.append(rule)
                     # pattern line
                     cur_pats.append(line[1:])
                     continue
                 mo = re.match(PAT_RULE_BODY_LINE, line)
                 if mo is not None:
                     grouplist, typename, featureassignments = mo.groups()
-                    cur_acts.append((grouplist, typename, featureassignments))
+                    # now try to match the feature assignments as the initializer of a dict, if that works,
+                    # it is a proper line, otherwise we get an error
+                    try:
+                        features = eval("dict("+featureassignments+")", GLOBALS)
+                    except Exception as ex:
+                        raise Exception(f"Not a valid feature assignment: {featureassignments}", ex)
+                    cur_acts.append((grouplist, typename, features))
                     continue
                 mo = re.fullmatch(PAT_MACRO_LINE, line)
                 if mo is not None:
@@ -253,105 +274,197 @@ class StringRegexAnnotator(StringGazetteerBase):
             # end for line
             if cur_acts:
                 # finish and add rule
-                rule = make_rule(cur_pats, cur_acts, cur_substs, backend=self.engine)
-                self.add(rule, list_features=list_features)
-                pass
+                rule = self.make_rule(cur_pats, cur_acts, substs=cur_substs, list_features=list_features)
+                self.rules.append(rule)
             else:
                 # if there was no last rule, there was no rule at all, this is invalid
                 raise Exception("No complete rule found")
 
+    def match_next(self, pat: Any, text: str,
+                   from_offset: int = 0,
+                   add_offset: int = 0,
+                   start_offsets: Union[None, list, set] = None,
+                   end_offsets: Union[None, list, set] = None,
+                   split_offsets: Union[None, list, set] = None
+                   ):
+        """
+        Find the next metch for the compiled pattern in text, at or after from_offset, but only if
+        all of the start/end/split offset limitations are satisfied (if present).
+
+        The from_offset is the offset within text, while add_offset is an additional offset value
+        to get added so we can compare to any of the start_offsets, end_offsets, split_offsets
+        (because text may just be a substring of the full text to which those offsets refer).
+
+        Args:
+            pat: a pre-compiled re/regex pattern
+            text: the text in which to search for the pattern
+            from_offset: the offset in text from which on to search for the pattern
+            add_offset: this gets added to a match offset in order to be comparable to the start/end/split offsets
+            start_offsets: a set/list of offsets where a match must start
+            end_offsets: a set/list of offsets where a match must end
+            split_offsets: a set/list of offsets where a match cannot cross
+
+        Returns:
+            None if no match is found, otherwise a tuple (start, end, groups) where groups is a tuple with all
+            groups from the RE, starting with group(0), then group(1) etc.
+        """
+        m = self.re.search(pat, text[from_offset:])
+        while m:
+            # in this loop we return the first match that is valid, iterating until we find one or
+            # no more matches are found
+            groups = [m.group(0)]
+            groups.extend(m.groups())
+            start, end = [o+from_offset for o in m.span()]
+
+            ostart = start + add_offset
+            oend = end + add_offset
+            if start_offsets and ostart not in start_offsets:
+                continue
+            if end_offsets and oend not in end_offsets:
+                continue
+            if split_offsets:
+                for i in range(ostart, oend):
+                    if i in split_offsets:
+                        continue
+            # the match should be valid, return it
+            return start, end, groups
+        # end while
+        return None
+
     def find_all(self, text: str,
+                 add_offset: int = 0,
                  longest_only: Union[None, bool] = None,
                  skip_longest: Union[None, bool] = None,
+                 select_rules: Union[None, str] = None,
                  start_offsets: Union[List, Set, None] = None,
                  end_offsets: Union[List, Set, None] = None,
-                 ws_offsets: Union[List, Set, None] = None,
                  split_offsets: Union[List, Set, None] = None):
         # first of all create a list of match iterator generators that correspond to each of the rules
-        BEYOND = len(text)+1
-        if self.engine == "regex":
-            # note: this should have been imported in the constructor, if needed
-            module = regex   # noqa: F821
-        else:
-            module = re
-        gens = [module.finditer(rule[0], text) for rule in self.rules]
-        # create a list of matches or None for each of the generators
-        matches = [one_or(gen) for gen in gens]
+        if longest_only is None:
+            longest_only = self.longest_only
+        if skip_longest is None:
+            skip_longest = self.skip_longest
+        if select_rules is None:
+            select_rules = self.select_rules
+        beyond = len(text)+1
+
+        # initialize the matches
+        matches = [self.match_next(rule.pattern, text, from_offset=0, add_offset=add_offset,
+                                   start_offsets=start_offsets, end_offsets=end_offsets,
+                                   split_offsets=split_offsets) for rule in self.rules]
 
         curoff = 0
         result = []
-        while curoff < len(text):
-            # go through all matches (or None) in rule order: ignore the Nones these are done
-            # if we got a match that is smaller than where we already are, consume until we get one at the curoff
-            # or beyond. Every time we got a match at curoff or beyond, check which offset is the smallest. Also
-            # update the current longest match length.
-            # NOTE: if smallestoff is still len(text)+1 after this, we did not find any match!
+        finished = False   # set to true once all matches are None, i.e. there is no match for any pattern left
+        while not finished and curoff < len(text):
             longestspan = 0
-            smallestoff = BEYOND
+            smallestoff = beyond
+            # find the smallest offset of all matches, and also the length of the longest span among all smallestoff
+            # matches
             for idx, match in enumerate(matches):
-                if match:
-                    if match.span[0] < curoff:
-                        match = one_or(gens[idx])
-                        while match and match.span[0] < curoff:
-                            match = one_or(gens[idx])
-                        matches[idx] = match
-                        if match is None:
-                            continue
-                    # we have a match for that rule
-                    if match.span[0] < smallestoff:
-                        # new smallest offset, also need to reset the longest match
-                        smallestoff = match.span[0]
-                        longestspan = 0
-                    mlen = match.span[1] - match.span[0]
-                    if mlen > longestspan:
-                        longestspan = mlen
-                # if match
+                # if the match is starting before curoff, update it
+                if match and match[0] < curoff:
+                    match = self.match_next(self.rules[idx].pattern, text,
+                                            from_offset=curoff, add_offset=add_offset,
+                                            start_offsets=start_offsets, end_offsets=end_offsets,
+                                            split_offsets=split_offsets)
+                    matches[idx] = match
+                if not match:
+                    continue
+                # if there actually is a valid match, use it to determine the smallest offset and the longest match len
+                if match[0] < smallestoff:
+                    # new smallest offset, also need to reset the longest match
+                    smallestoff = match[0]
+                    longestspan = 0
+                mlen = match[1] - match[0]
+                if mlen > longestspan:
+                    longestspan = mlen
             # for
-            if smallestoff == BEYOND:
+            # we now know where the next match(es) is/are happening and what the longest match(es) is/are
+            if smallestoff == beyond:
                 # no (more) matches found, break out of the while
                 break
+            curoff = smallestoff
             # we have at least one match still at smallestoff
             # depending on the strategy, select the rule to match:
             # all: all rules starting at smallestoff
             # first: the first rule at smallestoff
             # last: the last rule at smallestoff
             # firstlongest: the first rule at smallestoff which is of maximum length
+            # We select the indices of all rules for which the match should get considered
             idx2use = []
             lastidx = None
             for idx, match in enumerate(matches):
-                if match.span[0] != smallestoff:
+                if not match:
                     continue
-                if self.match == "all":
+                matchlen = match[1] - match[0]
+                if match[0] != smallestoff:
+                    continue
+                if not longest_only and select_rules == "all":
                     idx2use.append(idx)
-                elif self.match == "first":
+                elif longest_only and select_rules == "all" and matchlen == longestspan:
+                    idx2use.append(idx)
+                elif not longest_only and select_rules == "first":
                     idx2use.append(idx)
                     break
-                elif self.match == "firstlongest" and match.span[1]-match.span[0] == longestspan:
+                elif longest_only and select_rules == "first" and matchlen == longestspan:
                     idx2use.append(idx)
-                elif self.match == "last":
+                    break
+                elif not longest_only and select_rules == "last":
+                    lastidx = idx
+                elif longest_only and select_rules == "last" and matchlen == longestspan:
                     lastidx = idx
             # end for
-            if self.match == "last":
+            if select_rules == "last":
                 idx2use.append(lastidx)
             # now we have the list of idxs for which to add a match to the result
             for idx in idx2use:
                 match = matches[idx]
-                acts = self.rules[idx][1]
+                acts = self.rules[idx].actions
+                groups = match[2]
                 for act in acts:
-                    typename = act[0]
-                    # TODO: replace GroupNumber
-                    feats = act[1]
-                    toadd = Match(match.span[0], match.span[1], text[match.span[0]:match.span[1]], typename, feats)
-                    result.append(toadd)
+                    feats = replace_group(act.features, groups)
+                    for gn in act.groupnumbers:
+                        toadd = Match(start=match[0],
+                                      end=match[1],
+                                      match=groups[gn],
+                                      features=feats,
+                                      type=act.typename)
+                        result.append(toadd)
             # end for
+            # now depending on skip_longest, skip either one offset or the length of the longest match
+            if self.skip_longest:
+                curoff += longestspan
+            else:
+                curoff += 1
         # end while
         return result
 
-
     def __call__(self, doc: Document, **kwargs):
+        outset = doc.annset(self.outset_name)
+        annset = doc.annset(self.annset_name)
         chunks = []  # list of tuples (text, startoffset)
+        split_offsets = None
+        if self.split_type is not None:
+            split_offsets = set()
+            anns = annset.with_type(self.split_type)
+            for ann in anns:
+                for i in range(ann.start, ann.end):
+                    split_offsets.add(i)
+        start_offsets = None
+        end_offsets = None
+        if self.start_type is not None:
+            start_offsets = set()
+            anns = annset.with_type(self.start_type)
+            for ann in anns:
+                start_offsets.add(ann.start)
+        if self.end_type is not None:
+            end_offsets = set()
+            anns = annset.with_type(self.end_type)
+            for ann in anns:
+                end_offsets.add(ann.end)
         if self.containing_type is not None:
-            for ann in doc.annset(self.annset_name).with_type(self.containing_type):
+            for ann in annset.with_type(self.containing_type):
                 chunks.append((doc[ann], ann.start))
         else:
             chunks.append((doc.text, 0))
@@ -359,5 +472,8 @@ class StringRegexAnnotator(StringGazetteerBase):
             text = chunk[0]
             offset = chunk[1]
             # find the matches, add annotations, with the offsets adapted by offset
+            matches = self.find_all(text=text, add_offset=offset, start_offsets=start_offsets, end_offsets=end_offsets,
+                                    split_offsets=split_offsets)
+            for match in matches:
+                outset.add(match.start + offset, match.end + offset, match.type, match.feats)
         return doc
-